waste_sorter_hackathon/requirements.txt
ultralytics>=8.2.0
torch>=2.1.0
opencv-python>=4.9.0
PyYAML>=6.0.1
numpy>=1.24.0,<2.0
